fix add-friends-likes
